"#14 Fix missing link from last up block to Conv2D"
Include 'x_up' on the upsampling dense blocks allows the 'include_top'
layer to get the full feature map at the end of the upsampling branch.

This commit also changes the 'concatenate' inside __dense_block as the
previous method was failing to run on TensorFlow backend.

This was tested using Keras 2.0.2 and TensorFlow 1.1.0